Advanced/upt_intelligence.py
import logging
import os
import sqlite3
import json
import yaml
from typing import Dict, List
from datetime import datetime
from Advanced.protocol_llm import ProtocolLLM
from Advanced.protocol_knowledge_graph import ProtocolKnowledgeGraph
from Advanced.protocol_evolution_tracker import ProtocolEvolutionTracker

class ProtocolIntelligenceEngine:
    def __init__(self, config_path: str = "advanced_config.yaml", mock_db_path: str = "mock_data/mock_packets.db"):
        self.logger = logging.getLogger("UPTIntelligence")
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(levelname)s - %(message)s - %(pathname)s',
            filename='upt_intelligence.log',
            filemode='a'
        )
        self.config = self._load_config(config_path)
        self.mock_db_path = mock_db_path
        self.llm = ProtocolLLM()
        self.graph_db = ProtocolKnowledgeGraph()
        self.evolution_tracker = ProtocolEvolutionTracker()
        self._init_mock_db()
        self.logger.info("Initialized Protocol Intelligence Engine")

    def _load_config(self, config_path: str) -> Dict:
        """Load configuration from YAML"""
        try:
            with open(config_path, 'r') as f:
                return yaml.safe_load(f) or {}
        except Exception as e:
            self.logger.error(f"Failed to load config: {e}")
            return {
                'business_priorities': {'speed': 0.3, 'accuracy': 0.5, 'security': 0.2},
                'learning_rate': 0.1
            }

    def _init_mock_db(self):
        """Initialize mock database for testing"""
        try:
            conn = sqlite3.connect(self.mock_db_path, check_same_thread=False)
            cursor = conn.cursor()
            cursor.execute("""
                CREATE TABLE IF NOT EXISTS mock_packets (
                    id INTEGER PRIMARY KEY,
                    timestamp TEXT,
                    raw_packet BLOB,
                    protocol TEXT,
                    packet_length INTEGER
                )
            """)
            conn.commit()
            conn.close()
        except sqlite3.Error as e:
            self.logger.error(f"Failed to initialize mock DB: {e}")

    def deep_protocol_analysis(self, packet_data: bytes) -> Dict:
        """Perform deep analysis on packet data"""
        syntactic = self._analyze_syntax(packet_data)
        semantic = self._extract_semantics(packet_data)
        behavioral = self._predict_behavior(packet_data)
        security = self._assess_security(packet_data)
        return {
            "syntactic_analysis": syntactic,
            "semantic_analysis": semantic,
            "behavioral_prediction": behavioral,
            "security_implications": security
        }

    def _analyze_syntax(self, data: bytes) -> Dict:
        """Analyze packet syntax"""
        try:
            import numpy as np
            if not data:
                return {"entropy": 0.0, "patterns": [], "structure": {"header_length": 0, "payload_length": 0}, "encoding": "binary"}
            counts = np.bincount(list(data))
            probs = counts / len(data)
            probs = probs[probs > 0]  # Avoid log(0)
            entropy = -np.sum(probs * np.log2(probs))
            patterns = self._find_patterns(data)
            return {
                "entropy": float(entropy),
                "patterns": patterns,
                "structure": {"header_length": min(len(data), 20), "payload_length": max(0, len(data) - 20)},
                "encoding": "binary"
            }
        except Exception as e:
            self.logger.error(f"Syntax analysis error: {e}")
            return {"entropy": 0.0, "patterns": [], "structure": {"header_length": 0, "payload_length": 0}, "encoding": "binary"}

    def _find_patterns(self, data: bytes) -> List:
        """Find repeating patterns in data"""
        return [data[i:i+4].hex() for i in range(0, min(len(data), 20), 4) if data[i:i+4]]

    def _extract_semantics(self, data: bytes) -> Dict:
        """Extract semantic meaning using ProtocolLLM"""
        try:
            llm_result = self.llm.understand_protocol([data])
            return {
                "intent": llm_result.get("protocol_language", "unknown"),
                "payload_meaning": llm_result.get("semantic_embedding", []),
                "dialect": llm_result.get("protocol_language", "unknown")
            }
        except Exception as e:
            self.logger.error(f"Semantic analysis error: {e}")
            return {}

    def _predict_behavior(self, data: bytes) -> Dict:
        """Predict protocol behavior"""
        return self.evolution_tracker.track_protocol_evolution({datetime.now(): [data]})

    def _assess_security(self, data: bytes) -> Dict:
        """Assess security implications"""
        analysis = self._analyze_syntax(data)
        entropy = analysis.get("entropy", 0)
        return {
            "is_encrypted": entropy > 7.0,  # High entropy suggests encryption
            "risk_level": "high" if entropy > 7.5 else "low"
        }

    def handle_unknown_protocol(self, packet_data: bytes) -> Dict:
        """Handle unknown protocols with inference"""
        try:
            analysis = self.deep_protocol_analysis(packet_data)
            hypothesis = self.graph_db.infer_unknown_protocol(analysis)
            if hypothesis.get("confidence_scores", [0])[0] > 0.8:
                return {
                    "inferred_protocol": hypothesis["likely_protocols"][0],
                    "confidence": hypothesis["confidence_scores"][0],
                    "action": "translate"
                }
            else:
                new_protocol = self._create_protocol_definition(analysis)
                self.graph_db.add_protocol(new_protocol)
                return {
                    "inferred_protocol": new_protocol["name"],
                    "confidence": 0.5,
                    "action": "generic_translation"
                }
        except Exception as e:
            self.logger.error(f"Unknown protocol handling error: {e}")
            return {"inferred_protocol": "unknown", "confidence": 0.0, "action": "drop"}

    def _create_protocol_definition(self, analysis: Dict) -> Dict:
        """Create a new protocol definition"""
        return {
            "name": f"unknown_{hash(str(analysis))[:8]}",
            "features": analysis,
            "created_at": datetime.now().isoformat()
        }

    def analyze_encrypted_traffic(self, encrypted_stream: List[bytes]) -> Dict:
        """Analyze encrypted traffic patterns"""
        try:
            from sklearn.cluster import KMeans
            features = np.array([
                [len(p), self._analyze_syntax(p).get("entropy", 0)]
                for p in encrypted_stream
            ])
            if len(features) > 1:
                kmeans = KMeans(n_clusters=2, random_state=0).fit(features)
                labels = kmeans.labels_
                protocol_type = "TLS" if max(labels) > 0 else "unknown"
            else:
                protocol_type = "unknown"
            return {
                "likely_protocol": protocol_type,
                "confidence": 0.7,  # Placeholder
                "recommended_translation": self._suggest_translation(protocol_type)
            }
        except Exception as e:
            self.logger.error(f"Encrypted traffic analysis error: {e}")
            return {}

    def _suggest_translation(self, protocol_type: str) -> str:
        """Suggest translation target based on protocol type"""
        return "HTTP" if protocol_type != "unknown" else "JSON"

    def generate_translation_rule(self, source_proto: str, target_proto: str) -> Dict:
        """Generate a translation rule using LLM"""
        try:
            return {
                "rule": self.llm.generate_protocol_translator(source_proto, target_proto),
                "confidence": 0.8  # Placeholder for rule quality
            }
        except Exception as e:
            self.logger.error(f"Translation rule generation error: {e}")
            return {"rule": "def translate(data): return data", "confidence": 0.0}

    def apply_business_priorities(self, decision: Dict) -> Dict:
        """Adjust decisions based on business priorities"""
        try:
            priorities = self.config.get('business_priorities', {})
            estimated_latency = decision.get("estimated_latency_ms", 1.0)
            if estimated_latency <= 0:
                estimated_latency = 1.0
            decision["priority_score"] = (
                priorities.get("speed", 0.3) * (1 / estimated_latency) +
                priorities.get("accuracy", 0.5) * decision.get("confidence", 0.5) +
                priorities.get("security", 0.2) * (1 if not self._assess_security(decision.get("packet_data", b"")).get("is_encrypted", False) else 0.5)
            )
            return decision
        except Exception as e:
            self.logger.error(f"Business priorities error: {e}")
            decision["priority_score"] = 0.5
            return decision  