Advanced/requirements.txt
scikit-learn==1.2.2
networkx==3.1
pyyaml==6.0
numpy==1.23.5